refactor: 디렉토리 구조 변경에 따른 리팩토링 진행
디렉토리 구조에 따라 폴더 구조를 변경하였고, 추가적으로 Header위에 마진을 조정함
