+++ haul-fe/src/data/GlobalVariable.js
@@ -7,6 +7,7 @@ export const UrlMap = {
   choiceTranportTypeUrl: "/request/type",
   choiceDatePageUrl: "/request/date",
   choiceTimePageUrl: "/request/time",
+  choiceSrcPageUrl: "/request/source",
   morePageUrl: "/more",
   checkReservationPageUrl: "/check/list",
 };
