+++ haul-fe/src/routes/Router.jsx
@@ -1,4 +1,6 @@
 import { BrowserRouter, Outlet, Route, Routes } from "react-router-dom";
+import { useLocation } from "react-router-dom";
+import { useEffect } from "react";
 import TestView from "../views/TestView.jsx";
 import Splash from "../views/Splash/Splash.jsx";
 import Login from "../views/Login/Login.jsx";
@@ -7,14 +9,14 @@ import SignUP from "../views/SignUp/SignUp.jsx";
 import More from "../views/More/More.jsx";
 import Terms from "../views/Terms/Terms.jsx";
 import UserInfo from "../views/UserInfo/UserInfo.jsx";
-import ChoiceTransport from "../views/ChoiceTransport/ChoiceTransport.jsx";
-import ChoiceDate from "../views/ChoiceDate/ChoiceDate.jsx";
-import ChoiceTime from "../views/ChoiceTime/ChoiceTime";
-import { useLocation } from "react-router-dom";
-import { useEffect } from "react";
+import ChoiceTransport from "../views/Request/ChoiceTransport/ChoiceTransport.jsx";
+import ChoiceDate from "../views/Request/ChoiceDate/ChoiceDate.jsx";
+import ChoiceTime from "../views/Request/ChoiceTime/ChoiceTime";
+import ChoiceSrc from "../views/Request/ChoiceSrc/ChoiceSrc.jsx";
 import Contract from "../views/Contract/Contract.jsx";
 import Check from "../views/Check/Check.jsx";
 import UserPayments from "../views/UserPayments/UserPayments.jsx";
+
 
 const ScrollToTop = () => {
   const { pathname } = useLocation();
@@ -52,6 +54,7 @@ const Router = () => (
         <Route path="type" element={<ChoiceTransport />} />
         <Route path="date" element={<ChoiceDate />} />
         <Route path="time" element={<ChoiceTime />} />
+        <Route path="source" element={<ChoiceSrc />} />
       </Route>
       <Route path="/check" element={<Outlet />}>
         <Route path="list" element={<Check />} />
